drivers: timeaware_gpio: move under drivers/
This is generic enough to be a driver class and does not need to be
maintained under misc/.
