+++ drivers/Kconfig
@@ -90,6 +90,7 @@ source "drivers/smbus/Kconfig"
 source "drivers/spi/Kconfig"
 source "drivers/stepper/Kconfig"
 source "drivers/syscon/Kconfig"
+source "drivers/timeaware_gpio/Kconfig"
 source "drivers/timer/Kconfig"
 source "drivers/uaol/Kconfig"
 source "drivers/usb/Kconfig"
