+++ drivers/misc/Kconfig
@@ -25,7 +25,6 @@ source "drivers/misc/renesas_ra_external_interrupt/Kconfig"
 source "drivers/misc/renesas_rx_dtc/Kconfig"
 source "drivers/misc/renesas_rx_external_interrupt/Kconfig"
 source "drivers/misc/stm32n6_axisram/Kconfig"
-source "drivers/misc/timeaware_gpio/Kconfig"
 # zephyr-keep-sorted-stop
 
 endmenu
