+++ drivers/timeaware_gpio/Kconfig
@@ -21,6 +21,6 @@ config TIMEAWARE_GPIO_INIT_PRIORITY
 	help
 	  System initialization priority for timeaware GPIO drivers.
 
-source "drivers/misc/timeaware_gpio/Kconfig.timeaware_gpio_intel"
+source "drivers/timeaware_gpio/Kconfig.intel"
 
 endif # TIMEAWARE_GPIO
