+++ drivers/timeaware_gpio/Kconfig.intel
