+++ include/zephyr/drivers/timeaware_gpio.h
