+++ samples/drivers/timeaware_gpio/CMakeLists.txt
