+++ samples/drivers/timeaware_gpio/boards/intel_btl_s_crb.overlay
