+++ samples/drivers/timeaware_gpio/prj.conf
