+++ samples/drivers/timeaware_gpio/sample.yaml
